refactor(js): update types
