+++ packages/autocomplete-js/src/types/AutocompleteClassNames.ts
@@ -1,0 +1,17 @@
+export type AutocompleteClassNames = {
+  root: string;
+  form: string;
+  label: string;
+  inputWrapper: string;
+  input: string;
+  submitButton: string;
+  resetButton: string;
+  loadingIndicator: string;
+  panel: string;
+  panelLayout: string;
+  source: string;
+  sourceHeader: string;
+  list: string;
+  item: string;
+  sourceFooter: string;
+};
