+++ packages/autocomplete-js/src/types/AutocompleteCollection.ts
@@ -1,0 +1,8 @@
+import { BaseItem } from '@algolia/autocomplete-core';
+
+import { InternalAutocompleteSource } from './AutocompleteSource';
+
+export interface AutocompleteCollection<TItem extends BaseItem> {
+  source: InternalAutocompleteSource<TItem>;
+  items: TItem[];
+}
