+++ packages/autocomplete-js/src/types/AutocompleteDom.ts
@@ -1,0 +1,11 @@
+export type AutocompleteDom = {
+  inputWrapper: HTMLDivElement;
+  input: HTMLInputElement;
+  root: HTMLDivElement;
+  form: HTMLFormElement;
+  label: HTMLLabelElement;
+  submitButton: HTMLButtonElement;
+  resetButton: HTMLButtonElement;
+  loadingIndicator: HTMLDivElement;
+  panel: HTMLDivElement;
+};
